+++ packages/changelog-bot/package.json
@@ -6,6 +6,8 @@
     "@types/generate-changelog": "1.7.0",
     "@types/node": "10.3.1",
     "@wireapp/core": "2.9.2",
+    "chalk": "2.4.1",
+    "commander": "2.15.1",
     "generate-changelog": "1.7.1"
   },
   "description": "A changelog notification bot.",
